requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
jinja2
python-multipart
python-dotenv
google-genai